mm/gup: use new helpers for marking folios as referenced
Marking the folio as referenced in gup fast was first introduced by commit
8ee53820edfd ("thp: mmu_notifier_test_young"). Then later refractored
but still basically the samething in commit e93480537fd7 ("mm/gup: Mark all
pages PageReferenced in generic get_user_pages_fast()"). Some implementation
may changed over days by the goal is the same: provide hotness info for better
resource management. At the time it was introduced it was mainly
intended to be seen by the mmu notifier in khugepaged which checks the
folio's referenced flag.

We have made all users that cares about a folio's hotness info to the
new helper, so convert this updater too.

Using folio_mark_access here to allow promotion might be OK, but that
is slightly slower. To reduce overhead and keep the behaviour change
minimal, use the cautious helper to only mark the folio as referenced.

Signed-off-by: Kairui Song <kasong@tencent.com>